+++ app/Http/Controllers/Admin/AnswerCrudController.php
@@ -41,8 +41,8 @@ class AnswerCrudController extends CrudController
      */
     protected function setupListOperation()
     {
-        CRUD::setFromDb(); // columns
-
+        CRUD::addColumn('id');
+        CRUD::addColumn('answer');
         CRUD::removeColumn('question_id');
         CRUD::removeColumn('user_id');
 
@@ -64,13 +64,9 @@ class AnswerCrudController extends CrudController
     {
         CRUD::setValidation(AnswerRequest::class);
 
-//        CRUD::addFields(self::createFields());
+        CRUD::addFields(self::createFields());
 
-        /**
-         * Fields can be defined using the fluent syntax or array syntax:
-         * - CRUD::field('price')->type('number');
-         * - CRUD::addField(['name' => 'price', 'type' => 'number']));
-         */
+
     }
 
     /**
@@ -94,45 +90,38 @@ class AnswerCrudController extends CrudController
         return [
             [
                 'name' => 'answer',
-                'type' => 'textarea',
+                'type' => 'summernote',
                 'label' => 'Ҷавоб',
                 'attributes' => [
                     'placeholder' => 'Ҷавоб дар инҷо'
-                ]
+                ],
+                'options' => [
+                    'height' => 220,
+                ],
             ],
             [
-                'name' => 'image',
-                'type' => 'browse',
-                'label' => 'Акс',
-                'attributes' => [
-                    'placeholder' => 'Ҷавоб дар инҷо'
-                ]
-            ],
-            [
-                'type' => 'select2_from_ajax',
-                'name' => 'question_id',
+                'name' => 'question',
+                'type' => 'select2',
                 'label' => 'Ҷавоб ба саволи',
                 'attribute' => 'title',
-                'data_source' => url('/api/question/search'),
                 'placeholder'  => 'Ҷавоб диҳандаро интихоб кунед ...',
                 'minimum_input_length' => 1,
             ],
             [
+                'name' => 'parent',
+                'type' => 'select2',
+                'label' => 'Савол ба савол',
+                'attribute' => 'id',
+                'placeholder'  => 'Ҷавоб диҳандаро интихоб кунед ...',
+                'minimum_input_length' => 1,
+            ],
+
+            [
                 'type' => 'select2_from_ajax',
-                'name' => 'user_id',
-                'label' => 'Ҷавоб диҳанда (User)',
+                'name' => 'user',
+                'label' => 'Ҷавоб диҳанди',
                 'attribute' => 'name',
                 'entity'   => 'user',
-                'data_source' => url('/api/user/search'),
-                'placeholder'  => 'Ҷавоб диҳандаро интихоб кунед ...',
-                'minimum_input_length' => 1,
-            ],
-            [
-                'type' => 'select2_from_ajax',
-                'name' => 'parent_id',
-                'label' => 'Parent',
-                'attribute' => 'body',
-                'entity'   => 'parent',
                 'data_source' => url('/api/answer/search'),
                 'placeholder'  => 'Ҷавоб бар ҷавоб ...',
                 'minimum_input_length' => 1,