+++ app/Http/Controllers/HomeController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Article;
 use App\Models\Question;
 use Illuminate\Http\Request;
 
@@ -12,7 +13,9 @@ class HomeController extends Controller
     {
         $questions = Question::limit(10)->orderBy('created_at', 'desc')->get();
 
-        return view('welcome', compact('questions'));
+        $articles = Article::limit(6)->orderBy('created_at', 'desc')->get();
+
+        return view('welcome', compact('questions', 'articles'));
     }
 
 }